% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo3.R
\name{choose.data}
\alias{choose.data}
\title{Run shiny app to set up core image analysis parameters}
\usage{
choose.data()
}
\value{
a data frame
}
\description{
Run shiny app to set up core image analysis parameters
}
