% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo4.R
\name{choose.data3}
\alias{choose.data3}
\title{Run shiny app to set up core image analysis parameters}
\usage{
choose.data3()
}
\value{
a data frame
}
\description{
Run shiny app to set up core image analysis parameters
}
